mostly aesthetic changes i am noticing something peculiar with typescript in regards to inferring the unbound types consider the type of unbound.pipe
  type __Default<A, B> = unknown extends B ? A : B;

  pipe: <T, R>(
    thisArg: __Default<Object, ThisParameterType<(this: T, func: (me: T) => R) => R>>,
    ...args: Parameters<(this: T, func: (me: T) => R) => R>
    ): ReturnType<(this: T, func: (me: T) => R) => R>

called like

  pipe([1, 2, 3], func)

We basically want typescript to turn

  __Default<Object, ThisParameterType<(this: T, func: (me: T) => R) => R>>

into just

  T

However, I don't think it can to this until T is instantiated, since if T were to be
instantiated with a top type then we *would* result in thisArg: Object. Due to this,
no simplification can happen immediately.

Then, I notice that if func does not have a known type, it seems that typescript is not
able to instantiate T and then thisArg will end up getting assigned the type Object.

This is not desirable. We, the programmers, know that the type is supposed to be T, but
as a result of delegating this calculation to typescript at compile-time via __Default
and ThisParameterType, we have accidentally recruited some unintended side-effects.

The ideal solution would be to parse the type

  (this: T, func: (me: T) => R) => R

and then spit out the literal typscript code

  pipe: <T, R>(thisArg: T, func: (me: T) => R) => R

but I don't think such a thing is really plausible to do reasonably, especially since I
really don't want to have to try interfacing with the typescript compiler API again.

Perhaps there is a better way of implementing __Default?